# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API REST - Catégories, Produits et Acheteurs",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lister les catégories",
                "description": "Display a listing of the resource.",
                "operationId": "d753b5a85d077f3a311df9eb28de8a1a",
                "responses": {
                    "200": {
                        "description": "Liste des catégories"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Créer une catégorie",
                "description": "Store a newly created resource in storage.",
                "operationId": "1d6b893ef899e2aecece0a5e18049d8a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Café"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Une catégorie pour les produits de café"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie créée"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                }
            }
        },
        "/api/categories/{categorie}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Détails d'une catégorie avec ses produits",
                "description": "Display the specified resource.",
                "operationId": "204f07a46403659f3e16a6e06e95dcf1",
                "parameters": [
                    {
                        "name": "categorie",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie trouvée"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Mettre à jour une catégorie",
                "description": "Update the specified resource in storage.",
                "operationId": "2eaa317f2ea8871f48bcecdfe95e6a83",
                "parameters": [
                    {
                        "name": "categorie",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie mise à jour"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Supprimer une catégorie",
                "description": "Remove the specified resource from storage.",
                "operationId": "21702b2a8b7abdf7a40668fdcdf63a16",
                "parameters": [
                    {
                        "name": "categorie",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie supprimée"
                    }
                }
            }
        },
        "/api/produits": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Listes des produits",
                "description": "Display a listing of the resource.",
                "operationId": "c7e27af70c1d33e37f767417e0880e68",
                "responses": {
                    "200": {
                        "description": "Liste des produits"
                    }
                }
            },
            "post": {
                "tags": [
                    "Produits"
                ],
                "summary": "Créer un produit",
                "description": "Store a newly created resource in storage.",
                "operationId": "1dcf5bbc5058dd5e52384c289a96ba68",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prix",
                                    "stock",
                                    "categories_id"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prix": {
                                        "type": "numeric"
                                    },
                                    "stock": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "categories_id": {
                                        "type": "exists:categories,id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produit crée"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                }
            }
        },
        "/api/produits/{id}": {
            "get": {
                "tags": [
                    "Produits"
                ],
                "summary": "Detail d'un produit",
                "description": "Display the specified resource.",
                "operationId": "bd48b3dd26740d05293134d4155f41c1",
                "parameters": [
                    {
                        "name": "produit",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produit trouvé"
                    },
                    "404": {
                        "description": "Produit non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produits"
                ],
                "summary": "Mettre à jour un produit",
                "description": "Update the specified resource in storage.",
                "operationId": "db28bcb38f8bf8010cca584f0445eccd",
                "parameters": [
                    {
                        "name": "Produit",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string"
                                    },
                                    "prix": {
                                        "type": "numeric"
                                    },
                                    "stock": {
                                        "type": "integer"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "categories_id": {
                                        "type": "exists:categories,id"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produit mise à jour"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produits"
                ],
                "summary": "Supprimer un produit",
                "description": "Remove the specified resource from storage.",
                "operationId": "8766477c306a0545db7a9060ff0c6ca1",
                "parameters": [
                    {
                        "name": "produit",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produit supprimé"
                    }
                }
            }
        }
    },
...
}